Add branch documentation and organize/fix typos
