+++ src/components/actions/ChartData.js
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useState, useRef } from 'react'
-import ProgressMenu from '../ProgressMenu'
+import ProgressMenu from '../logic/ProgressMenu'
 import { Means } from '../Context'
 import Loading from '../views/Loading'
 import { rounded } from '../utility'
