+++ src/components/logic/Quiz.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef, useContext } from 'react'
-import Tally from './actions/Tally'
-import { Rounds } from './Context'
-import QuizQuestion from './views/QuizQuestion'
+import Tally from '../actions/Tally'
+import { Rounds } from '../Context'
+import QuizQuestion from '../views/QuizQuestion'
 
 
 export default function Quiz ({ data, round }) {
